changes ...
Ok we're reverting revision r409468 yet again.  Why?  Well we reverted it before
on commit r411263 and tried to merge changes for ldif stuff back into the 0.9.5
shared branch.  This failed miserably.

We then said ok we'll reinstate r409468 back on the release tag for ldif changes
and did that.  Then we deleted the branches/shared/0.9.5 branch and copied the 
releases/shared/0.9.5.1 modified release tag back to the 0.9.5 branch.



git-svn-id: https://svn.apache.org/repos/asf/directory/releases@411281 13f79535-47bb-0310-9956-ffa450edef68